chore: cleanup directory structure, UI components and migrations
- Add hardening_executions and default_admin_user migrations
- Add Input and Select UI components, refine Badge/Button/Stats
- Refine backend services, API handlers and existing migrations
- Update CI security workflow and Cargo manifests
- Ignore .claude/ local config

+++ backend-rust/src/services/compliance_engine.rs
@@ -32,12 +32,12 @@ pub struct ComplianceAssessment {
     pub status: Option<String>,
 }
 
-#[derive(Debug, Serialize, Deserialize)]
+#[derive(Debug, Serialize, Deserialize, sqlx::FromRow)]
 pub struct ComplianceOverview {
     pub target_id: Option<i32>,
     pub hostname: Option<String>,
     pub frameworks_assessed: Option<i64>,
-    pub avg_compliance_score: Option<BigDecimal>,
+    pub avg_compliance_score: Option<i32>,
     pub critical_violations: Option<i64>,
     pub high_violations: Option<i64>,
     pub last_assessment_date: Option<DateTime<Utc>>,
